Adds ch01_readme, container and fruit files:
container & fruit are the class files. These contain the main classes
and the sub-classes:
Fruit -> Apple | Orange
Container -> Barrel | Basket

Initial set-up of classes, generally work well but can be broken,
an interface will be needed to tie the two together as a composition,
and also to determin when a container is full and can be dispatched.

A mechanism will be also be needed to set and control the number of
fruits going into each corresponding container.